% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/BN-GWAS_applications_get_gene_gene_network_from_GTEx.R
\name{get_Bayesian_Network}
\alias{get_Bayesian_Network}
\title{Get Bayesian Network}
\usage{
get_Bayesian_Network(alpha, numCores, m.max, resid_mat, glasso_obj)
}
\arguments{
\item{alpha:}{p-value threshold used to define "causal" relations}

\item{numCores:}{number of cores for stable.fast algo in pcalg}

\item{m.max:}{max. number of conditioning variables}

\item{glasso_obj:}{glasso_obj with identified fixed gaps}
}
\value{
an PC object storing inferred causal relations between variables
}
\description{
Computes the Bayesian network for GTEx (tissue-specific), using Glasso output as pre-screening step 
Edges that are absent in glasso (ie partial correlation=0 after controlling for all other genes) 
are also set to be absent in pcalg fit
}
\examples{
get_Bayesian_Network(alpha, numCores, m.max, resid_mat, glasso_obj)

}
