% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/BN-GWAS_applications_get_gene_gene_network_from_GTEx.R
\name{get_Glasso_Graph}
\alias{get_Glasso_Graph}
\title{Learn Independence Graph}
\usage{
get_Glasso_Graph(resid_mat, n.cores)
}
\arguments{
\item{resid_mat}{adjusted tissue-specific RNA-seq expression profile}

\item{n.cores}{number of cores used}
}
\value{
glasso object with identified fixed gaps
}
\description{
Aim of this function is to employ Graphical lasso (via glassoFast) to learn an independence graph 
of the gene expressions by applying glasso, nodes that are connected means they are dependent 
when conditioned on ALL other variables in the graph
}
\examples{
get_Glasso_Graph(resid_mat, n.cores)

}
