% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BN_GWAS_application_merge_graphs.R
\name{Get_IDA_and_jointIDA}
\alias{Get_IDA_and_jointIDA}
\title{Get IDA and joinytIDA}
\usage{
Get_IDA_and_jointIDA(estimateGraph, cov_mat, target_genes)
}
\arguments{
\item{estimateGraph}{estimated adacency matrix for the whole causal graph including both "genes" and phenotype(s)}

\item{cov_mat}{the covariance matrix for the genes and phenotypes}

\item{target_genes}{target genes for causal effects estimation}
}
\value{
Effects_mat_final returned causal effects matrix with estimated causal effect
}
\description{
Estimate total and direct causal effects of genes on the outcome
}
\examples{
Get_IDA_and_jointIDA(estimateGraph, cov_mat, target_genes)

}
