% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BN_GWAS_application_merge_graphs.R
\name{get_iterative_pvalue_adjusted_cov}
\alias{get_iterative_pvalue_adjusted_cov}
\title{Iterative Pvalue-adjusted Covariance Matrix}
\usage{
get_iterative_pvalue_adjusted_cov(rho_pMax, cov, iterateNum, thres, n)
}
\arguments{
\item{rho_pMax}{m x m adjacency matrix for the gene-gene causal network: [i,j] =1 indicates an edge from i to j (graph package documentation)}

\item{cov}{calcuated covariance matrix between genes and outcome(s)}

\item{iterateNum}{number of iterations}

\item{thres}{threshold used to determine whether the penalization process is coverged!}

\item{n}{sample size of the pheno dataset used for covariance matirx calculation}
}
\value{
returned adjusted covariance matrix
}
\description{
get iterative pvalue-adjusted covariance matrix
}
\examples{
get_iterative_pvalue_adjusted_cov(rho_pMax, cov, iterateNum, thres)

}
