% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BN_GWAS_application_merge_graphs.R
\name{get_pvalue_adjusted_cov}
\alias{get_pvalue_adjusted_cov}
\title{Pvalue-adjusted Covariance Matrix}
\usage{
get_pvalue_adjusted_cov(rho_pMax, cov)
}
\arguments{
\item{rho_pMax}{m x m adjacency matrix for the gene-gene causal network: [i,j] =1 indicates an edge from i to j (graph package documentation)}

\item{cov}{calcuated covariance matrix between genes and outcome(s)}
}
\value{
returned adjusted covariance matrix
}
\description{
Get pvalue-adjusted covariance matrix
}
\examples{
get_pvalue_adjusted_cov(rho_pMax, cov)

}
