% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BN_GWAS_application_merge_graphs.R
\name{merge_causal_graphs}
\alias{merge_causal_graphs}
\title{Merge graph}
\usage{
merge_causal_graphs(gene_graph_adj, gene_to_outcome)
}
\arguments{
\item{gene_graph_adj}{m x m adjacency matrix for the gene-gene causal network: [i,j] =1 indicates an edge from i to j (graph package documentation)}

\item{gene_to_outcome}{m x 1 causal relations vector between genes and outcome: gene_to_outcome[i]=1 indicates causal connection from gene i to outcome}
}
\value{
(m+1) x (m+1) transposed adjacency matrix bgadd2[i,j]=1 indicates an edge from j to i in the causal graph
}
\description{
Merge gene-gene and gene-outcome graph into a whole graph
}
\examples{
merge_causal_graphs(gene_graph_adj, gene_to_outcome)

}
