% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/BN-GWAS_application_get_gene_outcome_causal_network.R
\name{PCSelect_Parallel}
\alias{PCSelect_Parallel}
\title{Find columns in dm, that have nonzero parcor with y given any other set of columns in dm}
\usage{
PCSelect_Parallel(
  y,
  dm,
  method = c("parallel"),
  mem.efficient = FALSE,
  num_workers,
  alpha,
  corMat = NA,
  corMethod = "standard",
  max_ord = 5,
  verbose = FALSE,
  directed = FALSE
)
}
\arguments{
\item{y}{Response Vector (length(y)=nrow(dm))}

\item{dm}{Data matrix (rows: samples, cols: nodes)}

\item{method}{Character string specifying method; the default, "parallel" provides an parallelised method to implement all the conditional independence tests.}

\item{mem.efficient}{If TRUE, uses less amount of memory at any time point while running the algorithm}

\item{num_workers}{The numbers of cores CPU to run the algorithm}

\item{alpha}{Significance level of individual partial correlation tests}

\item{corMat}{Correlation matrix of the input variables}

\item{corMethod}{"standard" or "Qn" for standard or robust correlation estimation}

\item{max_ord}{Maximum number of variables for the conditional set}

\item{verbose}{0-no output, 1-small output, 2-details}

\item{directed}{Logical; should the output graph be directed?}
}
\value{
a list containing two object
     G: boolean vector with connected nodes and zMin: Minimal z values
}
\description{
Find columns in dm, that have nonzero parcor with y given any other set of columns in dm
}
\examples{
PCSelect_Parallel(y, dm, method = c("parallel"), mem.efficient = FALSE,
                  num_workers, alpha, corMat = NA, corMethod = "standard", max_ord=5, verbose = FALSE,
                  directed = FALSE)
}
